backend: k8cache/k8cache.go: Add responseCapture struct for storing the response from K8s clusters
This help to capture the response that is coming from a K8s cluster
responseCapture struct has properties like statusCode, Header and Body.

statusCode: It is the status code (200, 404, 400 ) that will help to identify if the requests was successful  or not.
Headers: It is use to understand about the nature of the header that is received from the k8 clusters which can be
returned to the  client.
Body: Body is the value that will be stored in the cache ( key-value pair )