backend: k8cache/k8cache: Add UnmarshalCachedData to deserialize JSON string from cache
This adds a new function, UnmarshalCachedData which helps to deserialize a JSON
string received from cache. This helps to reconstructing the full HTTP
response (status, headers, body) when serving the K8's to the client.
this helps to provide the clarity about the incoming k8's responses to
the client.