backend: k8cache/k8cache: Add LoadFromCache to return data if exists in the cache
This add LoadFromCache to ensure that if the user has the permission to
view the resources then it will check if the generated key is found in
cache if the key is present in the cache then it will return directly to
the client in []byte.

This also adds unit test for LoadFromCache to ensure reliability of the
LoadFromCache.