backend: k8cache/k8cache: Add RequestK8ClusterAPIClusterAndStore to store the K8's API resposne in cache
This function adds RequestK8ClusterAPIAndStore ensures if the key was not found inside the cache
then this will make actual call to k8's and this will capture the response body and
convert the captured response to string. After converting it will store the response
with the key and TTL of 10*min.