backend: cmd/pagination: Add pagination implementation to return only page that was requested from the client.
This improves the pagination functionality, which help to return only
page that was requested from the client not the full list, this help
the frontend to fetch only page that the user wants to see the page not
the full list.

Currently the page size is limit =15, we can make it dynamic according
to the headlamp.