fontend: Loader: Fix to not freeze when main thread freezes
This avoids SVG that mui CircleProgress uses, so it doesn't freeze.

Also avoids flashing if the loading is finished before 0.25 seconds.

Fixes a11y issue to not use progressbar, because we always do not
know the progress, but instead are announcing a status.

Removes test which checks that people can add other props to it,
because no one is using this functionality.

Color and size props are used, and are still supported.