backend: k8cache/cacheStore: Add UnMarshalCacheData to deserialize JSON string
This adds UnmarshalCachedData deserialize a JSON string received from cache
back into a CachedResponseData struct. This function is used to reconstructing
the full HTTP response (status, headers, body) when serving the k8s to the client.
this is the essential part as it gives the clarity about the incoming k8s requests.