charts/headlamp: Fix for pluginctl without global install
Because there is no permission to do that now. So we run it without
installing it.

Also make sure the config and cache folders use somewhere writable.

The ConfigMap was rendering with extra newlines. At the top, and
after plugin.yml, which was causing an error when validating the
configuration inside the plugin manager code.