Add the concept of user-plugins
This will be used for the catalog-installed plugins, so we end up with
static, user, and dev plugins.
Static are plugins shipped in Headlamp.
User are plugins installed by the user (via the catalog).
Dev are plugins used during development, like when we are adding a new
feature to a plugin.