headlamp-plugin: Fix template/package-lock.json handling
There were some bugs in this, and now it generates a valid
template package-lock.json.

Except there is a circular dependency on the integrity field.
So we make sure we get the integrity field from the registry
and update the package-lock.json before npm ci is run.

If the package version is not found on the registry, then we just
do a npm install. This can happen in testing.